refactor(barriers): address code-review findings 1-9
Correctness:
- #1 validateMap now rejects a barrier that seals the goal off. cellGraph
  gains a HARD barrier-block path option; reachableFromEdge uses it, so a wall
  separating a start edge from the goal reads as unreachable at the validation
  boundary (not only in the offline playability sim).
- #2 mapHasContent() counts barriers, so barrier-only edits aren't silently
  wiped by Rebuild / start-edge change without the confirm prompt.
- #3 afterHistoryChange() clears the barrier selection on undo/redo, so an
  index shift can't leave a later endpoint-drag editing the wrong barrier.
- #4 editor validateMap mirrors the server exactly (finite coords via isFinite,
  in-world bounds, known style) so preview-clean == submit-clean.

Cleanup / altitude:
- #5 extracted server/geometry.js (sideOf/segmentsCross) shared by engine +
  cellGraph; extracted slideAlongSegmentWalls so bounceOffStoneEdges and
  bounceOffBarriers share one slide implementation (stone seam re-validates via
  a hoisted stoneEdgeActive; barriers always active).
- #6 extracted client/scripts/barrierArt.js — one copy of the fence/concrete art
  + seed/rng, concatenated into both the play and create bundles (and both HTML
  pages) instead of duplicated verbatim in draw.js and create.js.
- #7 config barriers.styles is now the allowed-style-key list (the dead
  per-style color/thickness fields are gone; art is the shared module's job).

Efficiency:
- #8 in-game barriers bake once to an offscreen canvas (keyed by geometry+style,
  supersampled) and blit with a rotate each frame; cache resets on newMap.
- #9 getBarrierBlockedEdges caches the blocked-edge Set on the map object
  (immutable per round) instead of rebuilding a signature string every tick.

barriers-test gains [E] (validateMap rejects a goal sealed by a full-height
wall). build + smoke + unit + content + map-format + ai-bumper-cross all green.
