+++ client/create.html
@@ -814,6 +814,7 @@
   <script defer src="scripts/auth.js"></script>
   <!-- BUILD: bundle-start -->
   <script src="scripts/rhill-voronoi-core.js"></script>
+  <script src="scripts/barrierArt.js"></script>
   <script src="scripts/create.js"></script>
   <script src="scripts/editorTools.js"></script>
   <script src="scripts/osk.js"></script>
