+++ client/play.html
@@ -110,6 +110,7 @@
         <!-- ADS_CONFIG -->
         <!-- BUILD: bundle-start -->
         <script src="scripts/rhill-voronoi-core.js"></script>
+        <script src="scripts/barrierArt.js"></script>
         <script src="scripts/game.js"></script>
         <script src="scripts/perf.js"></script>
         <script src="scripts/metrics.js"></script>
